Merge with Assaf Raman's Direct3D 10 render system.
This commit is the result of an effort made by Assaf Raman to build a basic D3D10 render system. Assaf has created a working prototype of the render system. After Assaf got his render system to render, Sinbad suggested that Assaf will merge his code with the code in the OGRE cvs head. This commit is the result of this merge.

Here are the main features of this commit:
1. The render system renders some of the demos.
2. A base (that will be improved in the future) for the fixedFunc.
3. A render system option to define that exceptions will be raised based on the severity of D3D10 information queue messages.
4. The texture loading with freeimage has basic functionality.
5. Shader parameters can be passed.
6. Most of the texture and rasterizer state functions are enabled.
7. The base for GPU memory buffers update is working.

Main features that are still missing:
1. A better fixedFunc emulation.
2. Shadow support.
3. Full texture support.
4. Render to texture.
5. Other various bugs and missing implementations.